+++ packages/ast-utilities/src/javascript/index.ts
@@ -1,10 +1,11 @@
-export {default as addInterface} from './addInterface';
+export {default as addComponentProps} from './addComponentProps';
 export {default as addImportSpecifier} from './addImportSpecifier';
 export {default as addImportStatement} from './addImportStatement';
-export {default as addComponentProps} from './addComponentProps';
+export {default as addInterface} from './addInterface';
+export {default as addVariableDeclaratorProps} from './addVariableDeclaratorProps';
 export {default as replaceJsxBody} from './replaceJsxBody';
 export {default as replaceStrings} from './replaceStrings';
-export {default as addVariableDeclaratorProps} from './addVariableDeclaratorProps';
+export {default as wrapJsxChildren} from './wrapJsxChildren';
 export {transform, transformSync} from './transform';
 export {parse, parseSync} from './parse';
 export {compose, astFrom} from './utilities';