+++ packages/ast-utilities/src/javascript/replaceJsxBody/replaceJsxBody.ts
@@ -4,26 +4,37 @@ import * as t from '@babel/types';
 import {astFrom} from '../utilities';
 
 export default function replaceJSXBody(parent: string, children: string) {
-  const parentJsx = astFrom(parent);
+  const newParentJsx = astFrom(parent);
 
   return {
     JSXElement(path: traverse.NodePath<t.JSXElement>) {
-      if (
-        t.isJSXIdentifier(path.node.openingElement.name) &&
-        path.node.openingElement.name.name === children &&
+      const oldParentName =
         t.isJSXElement(path.parent) &&
         t.isJSXIdentifier(path.parent.openingElement.name) &&
-        t.isExpressionStatement(parentJsx) &&
-        t.isJSXElement(parentJsx.expression) &&
-        t.isJSXIdentifier(parentJsx.expression.openingElement.name) &&
-        path.parent.openingElement.name.name !==
-          parentJsx.expression.openingElement.name.name
-      ) {
-        if (t.isJSXElement(parentJsx.expression)) {
-          const node = parentJsx.expression;
-          node.children = [path.node];
+        path.parent.openingElement.name.name;
 
-          path.replaceWith(node as any);
+      const newParentName =
+        t.isExpressionStatement(newParentJsx) &&
+        t.isJSXElement(newParentJsx.expression) &&
+        t.isJSXIdentifier(newParentJsx.expression.openingElement.name) &&
+        newParentJsx.expression.openingElement.name.name;
+
+      const nodeName =
+        t.isJSXIdentifier(path.node.openingElement.name) &&
+        path.node.openingElement.name.name;
+
+      if (
+        nodeName === children &&
+        newParentName !== nodeName &&
+        (!oldParentName || newParentName !== oldParentName)
+      ) {
+        if (
+          t.isExpressionStatement(newParentJsx) &&
+          t.isJSXElement(newParentJsx.expression)
+        ) {
+          const newNode = newParentJsx.expression;
+          newNode.children = [path.node];
+          path.replaceWith(newNode);
         }
       }
     },