+++ packages/ast-utilities/src/javascript/wrapJsxChildren/index.ts
@@ -1,0 +1,3 @@
+import wrapJsxChildren from './wrapJsxChildren';
+
+export default wrapJsxChildren;